Add: ✨ Logout the user
The main idea is when using the authentication token approach
was that the server does not store any information about the logged in client.
We don't need to send any requests, because firebase doesn't know and
doesn't care that we are, or were logged in.

The only thing that we need to change when log out is our state by using
our context API